refactor(fe): rename ClientIcon to UserIcon
