+++ front-end/src/lib/icons/UserIcon.svelte
