+++ front-end/src/lib/icons/index.ts
@@ -4,5 +4,5 @@ export { default as ProfileIcon } from './ProfileIcon.svelte'
 export { default as IndexIcon } from './IndexIcon.svelte'
 export { default as Oval } from './Oval.svelte'
 export { default as AnchorIcon } from './AnchorIcon.svelte'
-export { default as ClientIcon } from './ClientIcon.svelte'
+export { default as UserIcon } from './UserIcon.svelte'
 export { default as VolunteerIcon } from './VolunteerIcon.svelte'
